+++ src/__playground__/index.ts
@@ -1,3 +1,4 @@
+import { Readable } from 'stream';
 import { Messenger } from '../index.js';
 import { api } from './worker.js';
 
@@ -6,3 +7,14 @@ const messenger = new Messenger('channel');
 const service = await api.launchService({
     messengers: [messenger],
 });
+
+const arr = ['hello', 'world', 'foo', 'bar'];
+const myStream = new Readable({
+    read() {
+        if (!arr.length) this.push(null);
+
+        this.push(arr.splice(0, 1)[0]);
+    },
+});
+
+myStream.pipe(await messenger.createStream());